Add permission service and user show sad path
Permissions now set for users and non-users. If a non-user attempts to navigate the app, they will be redirected back to the home page and prompted to signin.

If users attempt to navigate to a non-existent user show page, they will now be redirected to their show page with a flash message letting them know that the user they tried to visit does not exist.